show appropriate message if the directory doesn't exist and cron job failed to execute

+++ lib/tasks/spotlight.rake
@@ -6,29 +6,33 @@ namespace :spotlight do
     task clear_cache: :environment do
       # Get the size of the tmp/network_files directory
       size = `du -sh tmp/network_files`.split("\t").first
-      Rails.logger.info("Beginning spotlight:riiif:clear_cache rake task." \
-        " Size of tmp/network_files/ is currently #{size}.")
-      # If the size is over 50GB, delete the oldest files until under 50GB
-      deleted_file_count = 0
-      while !size.nil? && size.match(/^(\d+(?:\.\d+)?)G$/) && Regexp.last_match(1).to_f > 50
-        # Get the oldest file in the tmp/network_files directory
-        oldest_file = `ls -t tmp/network_files | tail -n 1`.chomp
-        # Delete the oldest file
-        `rm tmp/network_files/#{oldest_file}`
+      if $?.exitstatus == 0
+        Rails.logger.info("Beginning spotlight:riiif:clear_cache rake task." \
+          " Size of tmp/network_files/ is currently #{size}.")
+        # If the size is over 50GB, delete the oldest files until under 50GB
+        deleted_file_count = 0
+        while !size.nil? && size.match(/^(\d+(?:\.\d+)?)G$/) && Regexp.last_match(1).to_f > 50
+          # Get the oldest file in the tmp/network_files directory
+          oldest_file = `ls -t tmp/network_files | tail -n 1`.chomp
+          # Delete the oldest file
+          `rm tmp/network_files/#{oldest_file}`
 
-        if $CHILD_STATUS.success?
-          deleted_file_count += 1
-          # Get the new size of the tmp/network_files directory
-          size = `du -sh tmp/network_files`.split("\t").first
-        else
-          # If removing the oldest file fails, there's something unexpected in the directory
-          Rails.logger.error("Failed to delete tmp/network_files/#{oldest_file} while running riiif:clear_cache rake task.")
-          break
+          if $CHILD_STATUS.success?
+            deleted_file_count += 1
+            # Get the new size of the tmp/network_files directory
+            size = `du -sh tmp/network_files`.split("\t").first
+          else
+            # If removing the oldest file fails, there's something unexpected in the directory
+            Rails.logger.error("Failed to delete tmp/network_files/#{oldest_file} while running riiif:clear_cache rake task.")
+            break
+          end
         end
+        Rails.logger.info("Completed spotlight:riiif:clear_cache rake task." \
+          " Size of tmp/network_files/ is now #{size}." \
+          " Cleared #{deleted_file_count} files.")
+      else
+        Rails.logger.error("Error running spotlight:riiif:clear_cache rake task. Unable to get size of tmp/network_files.")
       end
-      Rails.logger.info("Completed spotlight:riiif:clear_cache rake task." \
-        " Size of tmp/network_files/ is now #{size}." \
-        " Cleared #{deleted_file_count} files.")
     end
   end
